Refactor rte_triggers functions to static methods
